Update README.md
Friends is relatable because, the absence of a major character makes this program beautiful. Each member of the group has a unique narrative, and we get to share in each one. Each character's tale is as significant to the next as we follow them through their highs and lows as they go to work, home, and Central Perk.